# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Perpl Exchange stats API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PerplStats {
  dailyVolumeUsd: number;
  totalVolumeUsd: number;
  dailyFeesUsd: number;
  totalFeesUsd: number;
  dailyTradeCount: number;
  totalTradeCount: number;
  lastUpdatedMs: number;
  indexerHeadBlock: number;
  indexerTailBlock: number;
  chainHeadBlock: number;
  contractStartBlock: number;
  /** Fraction (0..1) of contract history indexed */
  indexedFraction: number;
  /** External baseline cumulative volume snapshot (e.g. from DefiLlama) */
  baselineVolumeUsd?: number;
  /** Unix ms timestamp of the baseline snapshot */
  baselineAtMs?: number;
  /** Volume our indexer has captured strictly after baselineAtMs */
  indexedDeltaVolumeUsd?: number;
  /** Distinct on-chain accounts seen since the contract deploy */
  totalUsers?: number;
  /** Latest aggregated open interest across all markets, in USD */
  openInterestUsd?: number;
  /** Unix ms when the OI snapshot was taken (0 if no snapshot yet) */
  openInterestAtMs?: number;
}

export interface OiPoint {
  timestampMs: number;
  oiUsd: number;
}

export interface MarketOi {
  perpId: number;
  symbol: string;
  oiUsd: number;
  markPrice: number;
}

export interface MarketStat {
  perpId: number;
  symbol: string;
  dailyVolumeUsd: number;
  dailyTradeCount: number;
}

export interface LeaderboardEntry {
  rank: number;
  /** Perpl internal account ID (uint256 as decimal string) */
  accountId: string;
  volumeUsd: number;
  feesUsd: number;
  /** Realized PnL aggregated from signed amount field on fills (USD) */
  pnlUsd: number;
  tradeCount: number;
}

export interface VolumePoint {
  timestampMs: number;
  volumeUsd: number;
  tradeCount: number;
}

export type GetMarketStats200 = {
  markets: MarketStat[];
};

export type GetLeaderboardParams = {
  period?: GetLeaderboardPeriod;
  metric?: GetLeaderboardMetric;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};

export type GetLeaderboardPeriod =
  (typeof GetLeaderboardPeriod)[keyof typeof GetLeaderboardPeriod];

export const GetLeaderboardPeriod = {
  day: "day",
  all: "all",
} as const;

export type GetLeaderboardMetric =
  (typeof GetLeaderboardMetric)[keyof typeof GetLeaderboardMetric];

export const GetLeaderboardMetric = {
  volume: "volume",
  pnl: "pnl",
} as const;

export type GetLeaderboard200Period =
  (typeof GetLeaderboard200Period)[keyof typeof GetLeaderboard200Period];

export const GetLeaderboard200Period = {
  day: "day",
  all: "all",
} as const;

export type GetLeaderboard200 = {
  period: GetLeaderboard200Period;
  entries: LeaderboardEntry[];
};

export type GetOiHistory200 = {
  points: OiPoint[];
  perMarket: MarketOi[];
};

export type GetVolumeTimeseries200 = {
  points: VolumePoint[];
};