# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Perpl Exchange stats API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [stats]
      summary: Top trading accounts ranked by volume
      parameters:
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [day, all]
            default: day
        - name: metric
          in: query
          required: false
          schema:
            type: string
            enum: [volume, pnl]
            default: volume
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: object
                required: [period, entries]
                properties:
                  period:
                    type: string
                    enum: [day, all]
                  entries:
                    type: array
                    items:
                      $ref: "#/components/schemas/LeaderboardEntry"
...
components:
  schemas:
...
    LeaderboardEntry:
      type: object
      required: [rank, accountId, volumeUsd, feesUsd, pnlUsd, tradeCount]
      properties:
        rank:
          type: integer
        accountId:
          type: string
          description: Perpl internal account ID (uint256 as decimal string)
        volumeUsd:
          type: number
        feesUsd:
          type: number
        pnlUsd:
          type: number
          description: Realized PnL aggregated from signed amount field on fills (USD)
        tradeCount:
          type: integer
...